DjangoGather/WebServer/服务器后端demo/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
Django==5.0.7
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ninja==1.3.0
django-redis==5.4.0
django-timezone-field==7.1
et_xmlfile==2.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
mssql-django==1.5
mysqlclient==2.2.7
openpyxl==3.1.5
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyodbc==5.2.0
pyOpenSSL==24.3.0
pypinyin==0.53.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
simplejwt==2.0.1
six==1.17.0
sqlparse==0.5.2
Twisted==24.11.0
txaio==23.1.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
